vfio/pci: Rename vfio_pci_device to vfio_pci_core_device
This is a preparation patch for separating the vfio_pci driver to a
subsystem driver and a generic pci driver. This patch doesn't change any
logic.

The new vfio_pci_core_device structure will be the main structure of the
core driver and later on vfio_pci_device structure will be the main
structure of the generic vfio_pci driver.

issue: 2690413
Change-Id: Id2cdc301e1990dcf310eb5f05f190b2574298e48
Signed-off-by: Max Gurtovoy <mgurtovoy@nvidia.com>