cover-letter: Introduce vfio_pci_core subsystem
This series splits the vfio_pci driver into two parts, a PCI driver and a
subsystem driver that will also be library of code. The main PCI driver,
vfio_pci.ko, will remain as before and it will use the library module
vfio_pci_core.ko to help create the vfio_device.

This series is intended to solve the issues that were raised in the
previous attempts for extending vfio-pci for device specific
functionality:

1. https://lore.kernel.org/kvm/20200518024202.13996-1-yan.y.zhao@intel.com
   by Yan Zhao
2. https://lore.kernel.org/kvm/20210702095849.1610-1-shameerali.kolothum.thodi@huawei.com
   by Longfang Liu

Also to support proposed future changes to virtio and other common
protocols to support migration:

https://lists.oasis-open.org/archives/virtio-comment/202106/msg00044.html

This subsystem framework will also ease adding new device specific
functionality to VFIO devices in the future by allowing another module to
provide the pci_driver that can setup a number of details before
registering to the VFIO subsystem, such as injecting its own operations.

This series also extends the "driver_override" mechanism. A flag is added
for PCI drivers that will declare themselves as "driver_override" capable
which sends their match table to the modules.alias file but otherwise
leaves them outside of the normal driver core auto-binding world, like
vfio_pci.

In order to get the best match for "driver_override" drivers, one can
create a userspace program to inspect the modules.alias, an example can be
found at:

https://github.com/maxgurtovoy/linux_tools/blob/main/vfio/bind_vfio_pci_driver.py

Which finds the 'best match' according to a simple algorithm: "the driver
with the fewest '*' matches wins."

For example, the vfio-pci driver will match to any pci device. So it will
have the maximal '*' matches.

In case we are looking for a match to a mlx5 based device, we'll have a
match to vfio-pci.ko and mlx5-vfio-pci.ko. We'll prefer mlx5-vfio-pci.ko
since it will have less '*' matches (probably vendor and device IDs will
match). This will work in the future for NVMe/Virtio devices that can
match according to a class code or other criteria.

issue: 2690413
Change-Id: Ib491bbb79f168541238c86f2e3f4726793156e0d
Signed-off-by: Max Gurtovoy <mgurtovoy@nvidia.com>