KVM: arm64: Fix the problem of everything ends up in slow-path
We are now clearing the active_asid when the vcpu gets scheduled
out. This has an adverse effect that, it increases the probability
of tasks ends up in slow-path for vmid allocation.

To fix this Will D has suggested a new algorithm,

ROLLOVER

  * Take lock
  * Inc generation
    => This will force everybody down the slow path
  * Record active VMIDs
  * Broadcast TLBI
    => Only active VMIDs can be dirty
    => Reserve active VMIDs and mark as allocated

VCPU SCHED IN

  * Set active VMID
  * Check generation
  * If mismatch then:
        * Take lock
        * Try to match a reserved VMID
        * If no reserved VMID, allocate new

VCPU SCHED OUT

* Clear active VMID

Please refer,
https://lore.kernel.org/kvmarm/20210803153036.GA31125@willie-the-truck/

Also, get rid of the unused smp_processor_id().

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>