kvm: ARM64: ANother attempt at the algo...
Extended tests with vmid bit set to 4, failed since the last commit.

This is another take on the algo as suggested by Will, but still fails.

Not sure, atomic64_set/read actually helps here to synchronise the fast
path and a concurrent rollover.

Added few comments as well to help review..

Signed-off-by: Shameer Kolothum <shameerali.kolothum.thodi@huawei.com>