{topost} scsi: hisi_sas: Change the timeout value of phyup
The time of phyup not only depends on SAS controller, but also the disk
connected (for SATA disk in directly attached situation, phyup occurs only
if SAS controller receives initial D2H from SATA disk after oob ready).
For some SATA disks, it spends much time on getting the disk ready after
reset or power-on, and it costs more than 10s sometimes.
Before we set phyup timeout HISI_SAS_WAIT_PHYUP_TIMEOUT to 20s in
hisi_sas_phy_oob_ready(), but it seems still not enough. According to
the specfication of some SATA disks such as ST14000NM0018, the max time
from power-on to ready is 30s which is also satified with other types of
disk, so 30s is more reasonable.
Currently the timeout of phyup after reset is 2s which is not enough. We
also decide to set it as HISI_SAS_WAIT_PHYUP_TIMEOUT as there is a chain
of dependency in hisi_sas_control_phy() with sem hisi_hba->sem if setting
the timeout greater than HISI_SAS_WAIT_PHYUP_TIMEOUT.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Signed-off-by: John Garry <john.garry@huawei.com>