sbitmap: Spread sbitmap word allocation over NUMA nodes
Currently sbitmap words are allocated in a single array from the NUMA node
passed in sbitmap_init_node().

However often the sbitmap will be accessed by the CPUs in the system - for
example, when BLK_MQ_F_TAG_HCTX_SHARED is set for the blk-mq tagset.

This can lead to performance issues where all CPUs in the system are doing
cross-NUMA node accesses to read/set/unset sbitmap tags.

Improve this by spreading the word allocations across all NUMA nodes as
evenly as possible. We set the per-CPU hint to fall within range of words
allocated for the NUMA node to which that CPU belongs.

Known issues:
- sbitmap resize does not work well for this scheme
- Improve updating hint for sbitmap_get() failure and sbitmap_put() when
  hint is outside range of CPU's NUMA node words.

Signed-off-by: John Garry <john.garry@huawei.com>