scsi: hisi_sas: Fix memory ordering in hisi_sas_task_deliver()
The memories for the slot should be observed to be written prior to
observing the slot as ready.

Prior to commit 26fc0ea74fcb ("scsi: libsas: Drop SAS_TASK_AT_INITIATOR"),
we had a spin_lock() + spin_unlock() immediately before before marking the
slot as ready. The spin_unlock() - with release semantics - caused the slot
memory to be observed to be written.

Now that the spin_lock() + spin_unlock() is gone, use a smp_wmb().

Fixes: 26fc0ea74fcb ("scsi: libsas: Drop SAS_TASK_AT_INITIATOR")
Signed-off-by: John Garry <john.garry@huawei.com>