{topost} scsi: libsas: Resume sas host when reset or enable phy via sysfs
Currently if reset or enable phy via sysfs when controller is suspended,
those operations will be ignored as SAS_HA_REGISTERED is cleared. If RPM
is enabled then we may aggressively suspend automatically. In this case
it may be difficult to enable or reset phy via sysfs, so resume sas host
when reset or enable phy via sysfs.

Signed-off-by: Xiang Chen <chenxiang66@hisilicon.com>
Reviewed-by: John Garry <john.garry@huawei.com>