scsi: libsas drivers: Prepare for reserved commands
Set the various LLDD fields to be able to use reserved commands.

We reserve 2x tags for libsas reserved tag handling, which should be
enough.

For drivers who manage tags internally for assigning a tag to sas_tasks
which do not have a request associated, continue to use the same carve
out.

Signed-off-by: John Garry <john.garry@huawei.com>