libata/scsi: libsas: Add sas_queuecommand_internal()
Add a callback for sht reserved_queuecommand callback. We need to add
libata helper ata_is_scmd_ata_internal() as it is difficult to know
whether the request should be queued as an ATA internal command.

We will store the sas_task in the scsi_cmnd payload also in future.

Signed-off-by: John Garry <john.garry@huawei.com>