{topost} scsi: hisi_sas: Drain bcast events in hisi_sas_rescan_topology()
In resetting the controller, SATA devices may be lost.

The issue is that when we insert the bcast events to rescan the topology
in hisi_sas_rescan_topology(), when we subsequentally nexus reset the
SATA devices in hisi_sas_async_I_T_nexus_reset(), there is a small timing
window in which the remote phy is down and we process the bcast event
(meaning that the disk is lost).

Ensure that all bcast events are processed prior to the nexus reset to
close this window.

Signed-off-by: John Garry <john.garry@huawei.com>