scsi: hisi_sas: Don't send bcast events from HW during nexus HA reset
Remote devices may go missing from the per-device nexus reset part of the
HA nexus, i.e after the controller reset. This is because libsas may find
the devices to be gone as the phy may be temporarily down when processing
the bcast event generated from the nexus reset. Set a flag to filter out
bcast events during this time to stop the devices being lost.

Signed-off-by: John Garry <john.garry@huawei.com>