vfio/mlx5: Refactor total_length name and usage
On the source side flow, change the usage of total_length to keep only
the size of the image as returned from the FW and rename it to
image_length. This is eliminating the usage of total_length inside the
SAVE command.

This is a preparation step for the next patches from the series where
more than one image could be managed on the migration file.

Signed-off-by: Shay Drory <shayd@nvidia.com>
Signed-off-by: Yishai Hadas <yishaih@nvidia.com>