vfio/ap: validate iova during dma_unmap and trigger irq disable
vfio container registers .dma_unmap() callback after the device is opened.
So it's fine for mdev drivers to call unpin page without checking matched
iova.

Now with iommufd, an access ops with an unmap callback is registered
when the device is bound to iommufd which is before .open_device()
is called. So it's not correct to do in the old way.

The fix is to let vfio_ap_mdev_dma_unmap() validate that unmap request
matches with one or more of these stashed values before attempting unpins.

Currently, each mapped iova is stashed in its associated vfio_ap_queue;
Each stashed iova represents IRQ that was enabled for a queue. Therefore,
if a match is found, trigger IRQ disable for this queue to ensure that
underlying firmware will no longer try to use the associated pfn after
the page is unpinned. IRQ disable will also handle the associated unpin.

Cc: Tony Krowiak <akrowiak@linux.ibm.com>
Cc: Halil Pasic <pasic@linux.ibm.com>
Cc: Jason Herne <jjherne@linux.ibm.com>
Signed-off-by: Matthew Rosato <mjrosato@linux.ibm.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>