vfio: Make open_device exclusive between group path and device cdev path
This prevents userspace from opening a specific vfio_device in both
group path and device cdev path. It is achieved by adding a flag
in vfio_device and vfio_core_device. In the device cdev path, the flag
in vfio_core_device is set, after opening the device, this flag is
propagated to vfio_device. Hence to prevent another open including device
cdev path and group path. In group path, this flag is never set.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>