arm_mpam/resctrl: Skip online/offline a CPU if not part of a class' affinity
If a CPU isn't part of the class' affinity, skip onlining it. Otherwise, it
will try to allocate a domain for it that will fault.

An example is when there are 2 memory MSCs with affinities to the first
64 CPUs (32 CPUs for each memory MSC) and 1 L3 cache MSC with affinity
to the first 32 CPUs. When a call back to online CPU 63 for the second
memory MSC is triggered, the function will first check the caches, and
it will fault and return as CPU 63 is not part of the caches' affinity,
without trying to check for the memory MSC. Thus, only 1 memory MSC
will show up, even though there are 2 that could be managed.

Also skip when offlining a CPU that's not part of the class' affinity.

Signed-off-by: Hesham Almatary <hesham.almatary@huawei.com>