arm_mpam/resctrl: Drop the assumption a class has affinity to all CPUs
If there are 128 CPUs in the system, but there is one memory and/or
cache MSC with affinity to a subset of these CPUs (e.g., 32 CPUs),
caches/memories MSC won't be picked at all and nothing will appear
in resctrl to manage valid MSCs.

This commit drops the assumption that caches or memory MSCs have
affinitiy to the entire CPUs in the system.

Signed-off-by: Hesham Almatary <hesham.almatary@huawei.com>